pyproject.toml
[project]
name = "studioagent"
version = "0.1.0"
description = "Autonomous AI video editing agent powered by Gemini"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "python-multipart>=0.0.18",
    "google-genai>=1.67.0",
    "google-cloud-storage>=2.19.0",
    "moviepy>=2.1.0",
    "opencv-python-headless>=4.10.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "python-dotenv>=1.0.1",
    "aiofiles>=24.1.0",
    "websockets>=14.0",
    "httpx>=0.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.25.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
